pyproject.toml
[project]
name = "basecode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "easydict>=1.13",
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.55.2",
    "wandb>=0.21.1",
]
